requirements.txt
attrs
lxml
requests
